tweaks, rename to stor_
