+++ crates/nu-command/src/stor/mod.rs
@@ -3,7 +3,7 @@ mod delete;
 mod insert;
 mod open;
 mod reset;
-mod stor;
+mod stor_;
 mod update;
 
 pub use create::StorCreate;
@@ -11,5 +11,5 @@ pub use delete::StorDelete;
 pub use insert::StorInsert;
 pub use open::StorOpen;
 pub use reset::StorReset;
-pub use stor::Stor;
+pub use stor_::Stor;
 pub use update::StorUpdate;
